Ignore test file build
- Add jest coverage config
- Move test files location
