+++ package.json
@@ -13,18 +13,22 @@
     "react-dom": "17.0.2"
   },
   "devDependencies": {
+    "@babel/plugin-syntax-jsx": "^7.12.13",
+    "@testing-library/jest-dom": "^5.12.0",
+    "@testing-library/react": "^11.2.7",
     "@types/jest": "^26.0.23",
     "@types/node": "^15.0.3",
     "@types/react": "^17.0.5",
     "@typescript-eslint/eslint-plugin": "^4.23.0",
+    "babel-jest": "^26.6.3",
     "eslint": "^7.26.0",
     "eslint-config-airbnb-typescript": "^12.3.1",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.23.2",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^26.6.3",
     "node-mocks-http": "^1.10.1",
-    "ts-jest": "^26.5.6",
     "typescript": "^4.2.4"
   }
 }
